reorder the proof that partition is a good lens
After quite some fighting with two answers
http://tex.stackexchange.com/questions/6988/how-to-sort-an-alphanumeric-list
http://tex.stackexchange.com/questions/152727/ifnumcomp-ref
talking about how to sort the trivlist in the partition list so that it
comes in the right order, I realized that there are two trivlists and
what's really needed is to sort the order of the two trivlists and the
text around them, not the items within the trivlists. So much for that
idea!

So this implements a much simpler solution: hard-code the order based on
whether it's the dissertation or not.